feat(yolo-tracker): refactor team tests for CPU environments #10
- Simplify test_team.py to 8 fast tests (create_batches only)
- Add test_team_model.py with 15 model tests (skipped by default)
- Update test_team_prediction.py with 11 prediction tests using __new__()
- All tests use proper mocking to avoid model loading

Test structure:
- test_team.py: 8 tests (run by default)
- test_team_model.py: 15 tests (skip unless RUN_MODEL_TESTS=1)
- test_team_prediction.py: 11 tests (run by default)

Total: 34 tests, 22 run by default
